requirements.txt
alabaster==0.7.12
amqp==2.2.2
Babel==2.6.0
beautifulsoup4==4.6.3
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0
django-allauth==0.35.0
django-cors-headers==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.90.2
factory-boy==2.10.0
Faker==0.8.11
gunicorn==19.7.1
idna==2.6
imagesize==1.1.0
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.1.0
mock==2.0.0
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==18.0
pbr==5.1.1
psycopg2==2.7.3.2
pyactiveresource==2.1.2
Pygments==2.2.0
PyJWT==1.6.4
pyparsing==2.3.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
redgreenunittest==0.1.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.0.0
ShopifyAPI==3.0.0
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.1.0
text-unidecode==1.2
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
